chore: restructure folders for better organization
